+++ apps/web/store/scanner.store.ts
@@ -1,0 +1,75 @@
+import { create } from 'zustand';
+import { devtools } from 'zustand/middleware';
+import { DashboardState, IBorders, IUploadFile } from '../types';
+
+export const useScannerStore = create<DashboardState>()(
+  devtools(
+    (set) => ({
+      counter: 0,
+      images: [],
+      borders: [],
+      selectedImageId: undefined,
+      addImage: (img: Omit<IUploadFile, 'border'>) => {
+        set((state) => ({
+          images: [...state.images, { ...img, id: state.counter }],
+          counter: state.counter + 1,
+          borders: [
+            ...state.borders,
+            {
+              id: state.counter,
+              borders: {
+                topLeft: {
+                  x: 0,
+                  y: 0,
+                },
+                topRight: {
+                  x: 0,
+                  y: 0,
+                },
+                bottomLeft: {
+                  x: 0,
+                  y: 0,
+                },
+                bottomRight: {
+                  x: 0,
+                  y: 0,
+                },
+              },
+            },
+          ],
+        }));
+      },
+      updateBorders: (id: number, borders: IBorders) => {
+        set((state) => ({
+          borders: state.borders.map((border) => {
+            if (border.id === id) {
+              return {
+                id,
+                borders,
+              };
+            }
+            return border;
+          }),
+          images: state.images.map((image) => {
+            if (image.id === id) {
+              return {
+                ...image,
+                processed: true,
+              };
+            }
+            return image;
+          }),
+        }));
+      },
+      selectImageById: (id: number) => {
+        set(() => ({
+          selectedImageId: id,
+        }));
+      },
+    }),
+    {
+      name: 'dashboard-state',
+      trace: true,
+    }
+  )
+);
